done with lab2

+++ lab2/DebugExercise/DebugExercise1.java
@@ -5,8 +5,9 @@ package DebugExercise;
  * feature work.
  */
 public class DebugExercise1 {
-    public static int divideThenRound(int top, int bottom) {
-        int quotient = top / bottom;
+    public static int divideThenRound(float top, float bottom) {
+        //int quotient = top / bottom;
+        float quotient = top / bottom;
         int result = Math.round(quotient);
         return result;
     }
